requirements.txt
pysmb==1.2.8
python-dotenv==0.20.0
watchdog==2.1.9
